libYARP_math: move gsl dependent file in a separate directory
